requirements.txt
fastapi==0.103.1
uvicorn==0.22.0
psycopg2-binary==2.9.6
pydantic==1.10.9
pandas==1.5.3
matplotlib==3.7.1
